refactor: Update project files to remove netstandard2.0 support and adjust target frameworks feat: Add message scheduling capabilities with in-memory and storage-backed options chore: Clean up unused package references across multiple projects test: Update test projects to reflect changes in dependencies and namespaces
