+++ src/HeroMessaging.Observability.HealthChecks/HeroMessaging.Observability.HealthChecks.csproj
@@ -5,15 +5,8 @@
     <PackageTags>messaging;healthcheck;monitoring;diagnostics;observability;aspnetcore</PackageTags>
   </PropertyGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-		<PackageReference Include="Microsoft.Bcl.TimeProvider" Version="9.0.10" />
-		<PackageReference Include="System.Collections.Immutable" Version="9.0.10" />
-	</ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="9.0.10" />
-	  <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.10" />
-	  <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.10" />
   </ItemGroup>
 
   <ItemGroup>
