+++ src/HeroMessaging.Observability.OpenTelemetry/HeroMessaging.Observability.OpenTelemetry.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <!-- OpenTelemetry packages don't support netstandard2.0 with latest versions -->
-    <TargetFrameworks>net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <Description>OpenTelemetry observability plugin for HeroMessaging. Provides distributed tracing, metrics, and correlation for message processing, saga orchestration, and inbox/outbox patterns.</Description>
     <PackageTags>messaging;opentelemetry;observability;tracing;metrics;distributed-tracing</PackageTags>
   </PropertyGroup>
@@ -16,12 +16,6 @@
     <PackageReference Include="OpenTelemetry" Version="1.13.1" />
     <PackageReference Include="OpenTelemetry.Api" Version="1.13.1" />
     <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.13.1" />
-    <PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.13.1" />
-    <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.13.1" />
-    <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.13.0" />
-    <PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.13.0" />
-    <PackageReference Include="OpenTelemetry.Instrumentation.Runtime" Version="1.13.0" />
-    <PackageReference Include="System.Diagnostics.DiagnosticSource" Version="9.0.10" />
   </ItemGroup>
 
 </Project>
