+++ src/HeroMessaging.Security/HeroMessaging.Security.csproj
@@ -4,17 +4,4 @@
     <ProjectReference Include="..\HeroMessaging.Abstractions\HeroMessaging.Abstractions.csproj" />
   </ItemGroup>
 
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
-  </PropertyGroup>
-
-  <!-- AesGcm is not available in netstandard2.0, only .NET Core 3.0+ -->
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <PackageReference Include="System.Security.Cryptography.Algorithms" Version="4.3.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.10" />
-  </ItemGroup>
-
 </Project>
