+++ src/HeroMessaging.Serialization.Json/HeroMessaging.Serialization.Json.csproj
@@ -9,10 +9,6 @@
     <ProjectReference Include="..\HeroMessaging.Abstractions\HeroMessaging.Abstractions.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="System.Text.Json" Version="9.0.10" />
-  </ItemGroup>
-
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
     <Using Include="System" />
     <Using Include="System.Collections.Generic" />
