+++ src/HeroMessaging.Serialization.MessagePack/HeroMessaging.Serialization.MessagePack.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <!-- MessagePack v3.x doesn't support netstandard2.0 -->
-    <TargetFrameworks>net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <Description>High-performance MessagePack binary serialization plugin for HeroMessaging. Provides fast, compact message serialization with cross-platform support.</Description>
     <PackageTags>messaging;serialization;messagepack;binary;performance</PackageTags>
   </PropertyGroup>
