+++ src/HeroMessaging/Configuration/ExtensionsToScheduling.cs
@@ -3,12 +3,14 @@ using HeroMessaging.Scheduling;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace HeroMessaging.Abstractions.Configuration;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// Extension methods for configuring message scheduling.
 /// </summary>
-public static class SchedulingExtensions
+public static class ExtensionsToScheduling
 {
     /// <summary>
     /// Adds message scheduling capabilities to HeroMessaging.
