--- src/HeroMessaging/GlobalUsings.cs
@@ -1,10 +1,0 @@
-#if NETSTANDARD2_0
-global using System;
-global using System.Collections.Concurrent;
-global using System.Collections.Generic;
-global using System.IO;
-global using System.Linq;
-global using System.Text;
-global using System.Threading;
-global using System.Threading.Tasks;
-#endif
