+++ src/HeroMessaging/HeroMessaging.csproj
@@ -16,18 +16,11 @@
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
 		<PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
 		<PackageReference Include="System.Threading.Channels" Version="9.0.10" />
-		<PackageReference Include="System.Collections.Immutable" Version="9.0.10" />
-		<PackageReference Include="System.Text.Encodings.Web" Version="9.0.10" />
-		<PackageReference Include="System.Text.Json" Version="9.0.10" />
-		<PackageReference Include="System.Threading.Tasks.Extensions" Version="4.6.3" />
-		<PackageReference Include="Microsoft.Bcl.TimeProvider" Version="9.0.10" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.10" />
 		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.10" />
 		<PackageReference Include="Microsoft.Extensions.Options" Version="9.0.10" />
-		<PackageReference Include="System.Threading.Tasks.Dataflow" Version="9.0.10" />
 	</ItemGroup>
 
 	<ItemGroup>
