+++ tests/HeroMessaging.Observability.HealthChecks.Tests/HeroMessaging.Observability.HealthChecks.Tests.csproj
@@ -21,20 +21,12 @@
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 			<PrivateAssets>all</PrivateAssets>
 		</PackageReference>
-		<PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="9.0.10" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<ProjectReference Include="..\..\src\HeroMessaging.Observability.HealthChecks\HeroMessaging.Observability.HealthChecks.csproj" />
 		<ProjectReference Include="..\..\src\HeroMessaging.Abstractions\HeroMessaging.Abstractions.csproj" />
 		<ProjectReference Include="..\HeroMessaging.Tests\HeroMessaging.Tests.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Update="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
-		<PackageReference Update="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
-		<PackageReference Update="PolySharp" Version="1.15.0" />
-		<PackageReference Update="System.Collections.Immutable" Version="9.0.0" />
 	</ItemGroup>
 
 </Project>
