+++ tests/HeroMessaging.Security.Tests/Unit/AesGcmMessageEncryptorTests.cs
@@ -1,6 +1,7 @@
 using System.Security.Cryptography;
 using HeroMessaging.Abstractions.Security;
 using HeroMessaging.Security.Encryption;
+using Xunit;
 
 namespace HeroMessaging.Security.Tests.Unit;
