+++ tests/HeroMessaging.Security.Tests/Unit/ClaimsAuthenticationProviderTests.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using HeroMessaging.Abstractions.Security;
 using HeroMessaging.Security.Authentication;
+using Xunit;
 
 namespace HeroMessaging.Security.Tests.Unit;
