+++ tests/HeroMessaging.Security.Tests/Unit/HmacSha256MessageSignerTests.cs
@@ -1,6 +1,7 @@
 using System.Security.Cryptography;
 using HeroMessaging.Abstractions.Security;
 using HeroMessaging.Security.Signing;
+using Xunit;
 
 namespace HeroMessaging.Security.Tests.Unit;
