+++ tests/HeroMessaging.Security.Tests/Unit/PolicyAuthorizationProviderTests.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using HeroMessaging.Abstractions.Security;
 using HeroMessaging.Security.Authorization;
+using Xunit;
 
 namespace HeroMessaging.Security.Tests.Unit;
