+++ tests/HeroMessaging.Serialization.Protobuf.Tests/HeroMessaging.Serialization.Protobuf.Tests.csproj
@@ -29,10 +29,4 @@
 		<ProjectReference Include="..\HeroMessaging.Tests\HeroMessaging.Tests.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<PackageReference Update="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
-		<PackageReference Update="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
-		<PackageReference Update="System.Collections.Immutable" Version="9.0.0" />
-	</ItemGroup>
-
 </Project>
