+++ tests/HeroMessaging.Storage.PostgreSql.Tests/HeroMessaging.Storage.PostgreSql.Tests.csproj
@@ -32,10 +32,4 @@
 		<ProjectReference Include="..\HeroMessaging.Tests\HeroMessaging.Tests.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<PackageReference Update="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
-		<PackageReference Update="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
-		<PackageReference Update="System.Collections.Immutable" Version="9.0.0" />
-	</ItemGroup>
-
 </Project>
