+++ tests/HeroMessaging.Tests/HeroMessaging.Tests.csproj
@@ -22,7 +22,6 @@
 			<PrivateAssets>all</PrivateAssets>
 		</PackageReference>
 		<PackageReference Include="Moq" Version="4.20.72" />
-		<PackageReference Include="Newtonsoft.Json" Version="13.0.4" />
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.10" />
 		<PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="9.0.10" />
 		<PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.10" />
@@ -38,12 +37,6 @@
 		<ProjectReference Include="..\..\src\HeroMessaging.Abstractions\HeroMessaging.Abstractions.csproj" />
 		<ProjectReference Include="..\..\src\HeroMessaging.Observability.OpenTelemetry\HeroMessaging.Observability.OpenTelemetry.csproj" />
 		<ProjectReference Include="..\..\src\HeroMessaging.Observability.HealthChecks\HeroMessaging.Observability.HealthChecks.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Update="Microsoft.Extensions.DependencyInjection.Abstractions" Version="9.0.0" />
-		<PackageReference Update="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
-		<PackageReference Update="System.Collections.Immutable" Version="9.0.0" />
 	</ItemGroup>
 
 </Project>
