+++ tests/HeroMessaging.Tests/TestUtilities/Examples/EnhancedBuilderExample.cs
@@ -3,7 +3,7 @@ using HeroMessaging.Abstractions.Messages;
 using HeroMessaging.Abstractions.Sagas;
 using HeroMessaging.Orchestration;
 
-namespace HeroMessaging.Tests.TestUtilities;
+namespace HeroMessaging.Tests.TestUtilities.Examples;
 
 /// <summary>
 /// Example demonstrating the enhanced state machine builder API
